Replace the interface with a deprecation error
This will be released as a minor version once the migration from
livefyre to coral has finished.

It is largely required by the infographics team as they may not be able
to update all their projects. This will allow older projects to be
rebuilt with the latest dependencies and prevent unknown errors being
thrown.